bonzour

+++ srcs/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: adeletan <adeletan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/06 00:13:29 by adeletan          #+#    #+#             */
-/*   Updated: 2016/11/06 04:33:30 by adeletan         ###   ########.fr       */
+/*   Updated: 2016/11/17 17:22:44 by rpagot           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t		ft_strlen(const char *s)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s[i])
